+++ docs/news.rst
@@ -42,6 +42,7 @@ Psycopg 3.0.17 (unreleased)
 
 - Fix segfaults on fork on some Linux systems using `ctypes` implementation
   (:ticket:`#300`).
+- Load bytea as bytes, not memoryview, using `ctypes` implementation.
 
 
 Current release